add bitflags support to inputdispatcher
